requirements.txt
Flask
flask-mysqldb
qrcode
Pillow
python-dotenv
pytest